requirements.txt
python-magic
http://pefile.googlecode.com/files/pefile-1.2.10-139.tar.gz#egg=pefile
PrettyTable
pydeep
requests
SQLAlchemy
pycrypto
olefile
oletools
BeautifulSoup4
bottle
pylzma
pyelftools
bitstring
dnspython
git+https://github.com/smarnach/pyexiftool.git#egg=pyexiftool
pyasn1
M2Crypto
pymisp
django
pypssl
pypdns
scandir
pype32
pbkdf2
